melt.default {reshape2}	R Documentation

Melt a vector. For vectors, makes a column of a data frame

Description

Melt a vector. For vectors, makes a column of a data frame

Usage

  ## Default S3 method:
 melt(data, ..., na.rm = FALSE,
    value.name = "value")

Arguments

data
vector to melt

na.rm
Should NA values be removed from the data set? This will convert explicit missings to implicit missings.

...
further arguments passed to or from other methods.

value.name
name of variable used to store values

[Package reshape2 version 1.2.2 Index]