melt_check {reshape2}	R Documentation

Check that input variables to melt are appropriate.

Description

If id.vars or measure.vars are missing, melt_check will do its best to impute them. If you only supply one of id.vars and measure.vars, melt will assume the remainder of the variables in the data set belong to the other. If you supply neither, melt will assume discrete variables are id variables and all other are measured.

Usage

  melt_check(data, id.vars, measure.vars)

Arguments

data
data frame

id.vars
vector of identifying variable names or indexes

measure.vars
vector of Measured variable names or indexes

Value

a list giving id and measure variables names.

[Package reshape2 version 1.2.2 Index]