parse_formula {reshape2}	R Documentation

Parse casting formulae.

Description

There are a two ways to specify a casting formula: either as a string, or a list of quoted variables. This function converts the former to the latter.

Usage

parse_formula(formula = "...  ~ variable", varnames, value.var = "value")

Arguments

formula
formula to parse

varnames
names of all variables in data

value.var
name of variable containing values

Details

Casting formulas separate dimensions with ~ and variables within a dimension with + or *. . can be used as a placeholder, and ... represents all other variables not otherwise used.

Examples

reshape2:::parse_formula("a + ...", letters[1:6])
reshape2:::parse_formula("a ~ b + d")
reshape2:::parse_formula("a + b ~ c ~ .")

[Package reshape2 version 1.4.3 Index]