colsplit {reshape2}	R Documentation

Split a vector into multiple columns

Description

Useful for splitting variable names that a combination of multiple variables. Uses type.convert to convert each column to correct type, but will not convert character to factor.

Usage

colsplit(string, pattern, names)

Arguments

string
character vector or factor to split up

pattern
regular expression to split on

names
names for output columns

Examples

x <- c("a_1", "a_2", "b_2", "c_3")
vars <- colsplit(x, "_", c("trt", "time"))
vars
str(vars)

[Package reshape2 version 1.4.4 Index]