Melt a data frame into form suitable for easy casting.

You need to tell melt which of your variables are id variables, and which are measured variables. If you only supply one of id.vars and measure.vars, melt will assume the remainder of the variables in the data set belong to the other. If you supply neither, melt will assume factor and character variables are id variables, and all others are measured.

# S3 method for class 'data.frame'
melt(
  data,
  id.vars,
  measure.vars,
  variable.name = "variable",
  ...,
  na.rm = FALSE,
  value.name = "value",
  factorsAsStrings = TRUE
)

Arguments

data

data frame to melt

id.vars

vector of id variables. Can be integer (variable position) or string (variable name). If blank, will use all non-measured variables.

measure.vars

vector of measured variables. Can be integer (variable position) or string (variable name)If blank, will use all non id.vars

variable.name

name of variable used to store measured variable names

...

further arguments passed to or from other methods.

na.rm

Should NA values be removed from the data set? This will convert explicit missings to implicit missings.

value.name

name of variable used to store values

factorsAsStrings

Control whether factors are converted to character when melted as measure variables. When FALSE, coercion is forced if levels are not identical across the measure.vars.

See also

Examples

names(airquality) <- tolower(names(airquality))
melt(airquality, id=c("month", "day"))
#>     month day variable value
#> 1       5   1    ozone  41.0
#> 2       5   2    ozone  36.0
#> 3       5   3    ozone  12.0
#> 4       5   4    ozone  18.0
#> 5       5   5    ozone    NA
#> 6       5   6    ozone  28.0
#> 7       5   7    ozone  23.0
#> 8       5   8    ozone  19.0
#> 9       5   9    ozone   8.0
#> 10      5  10    ozone    NA
#> 11      5  11    ozone   7.0
#> 12      5  12    ozone  16.0
#> 13      5  13    ozone  11.0
#> 14      5  14    ozone  14.0
#> 15      5  15    ozone  18.0
#> 16      5  16    ozone  14.0
#> 17      5  17    ozone  34.0
#> 18      5  18    ozone   6.0
#> 19      5  19    ozone  30.0
#> 20      5  20    ozone  11.0
#> 21      5  21    ozone   1.0
#> 22      5  22    ozone  11.0
#> 23      5  23    ozone   4.0
#> 24      5  24    ozone  32.0
#> 25      5  25    ozone    NA
#> 26      5  26    ozone    NA
#> 27      5  27    ozone    NA
#> 28      5  28    ozone  23.0
#> 29      5  29    ozone  45.0
#> 30      5  30    ozone 115.0
#> 31      5  31    ozone  37.0
#> 32      6   1    ozone    NA
#> 33      6   2    ozone    NA
#> 34      6   3    ozone    NA
#> 35      6   4    ozone    NA
#> 36      6   5    ozone    NA
#> 37      6   6    ozone    NA
#> 38      6   7    ozone  29.0
#> 39      6   8    ozone    NA
#> 40      6   9    ozone  71.0
#> 41      6  10    ozone  39.0
#> 42      6  11    ozone    NA
#> 43      6  12    ozone    NA
#> 44      6  13    ozone  23.0
#> 45      6  14    ozone    NA
#> 46      6  15    ozone    NA
#> 47      6  16    ozone  21.0
#> 48      6  17    ozone  37.0
#> 49      6  18    ozone  20.0
#> 50      6  19    ozone  12.0
#> 51      6  20    ozone  13.0
#> 52      6  21    ozone    NA
#> 53      6  22    ozone    NA
#> 54      6  23    ozone    NA
#> 55      6  24    ozone    NA
#> 56      6  25    ozone    NA
#> 57      6  26    ozone    NA
#> 58      6  27    ozone    NA
#> 59      6  28    ozone    NA
#> 60      6  29    ozone    NA
#> 61      6  30    ozone    NA
#> 62      7   1    ozone 135.0
#> 63      7   2    ozone  49.0
#> 64      7   3    ozone  32.0
#> 65      7   4    ozone    NA
#> 66      7   5    ozone  64.0
#> 67      7   6    ozone  40.0
#> 68      7   7    ozone  77.0
#> 69      7   8    ozone  97.0
#> 70      7   9    ozone  97.0
#> 71      7  10    ozone  85.0
#> 72      7  11    ozone    NA
#> 73      7  12    ozone  10.0
#> 74      7  13    ozone  27.0
#> 75      7  14    ozone    NA
#> 76      7  15    ozone   7.0
#> 77      7  16    ozone  48.0
#> 78      7  17    ozone  35.0
#> 79      7  18    ozone  61.0
#> 80      7  19    ozone  79.0
#> 81      7  20    ozone  63.0
#> 82      7  21    ozone  16.0
#> 83      7  22    ozone    NA
#> 84      7  23    ozone    NA
#> 85      7  24    ozone  80.0
#> 86      7  25    ozone 108.0
#> 87      7  26    ozone  20.0
#> 88      7  27    ozone  52.0
#> 89      7  28    ozone  82.0
#> 90      7  29    ozone  50.0
#> 91      7  30    ozone  64.0
#> 92      7  31    ozone  59.0
#> 93      8   1    ozone  39.0
#> 94      8   2    ozone   9.0
#> 95      8   3    ozone  16.0
#> 96      8   4    ozone  78.0
#> 97      8   5    ozone  35.0
#> 98      8   6    ozone  66.0
#> 99      8   7    ozone 122.0
#> 100     8   8    ozone  89.0
#> 101     8   9    ozone 110.0
#> 102     8  10    ozone    NA
#> 103     8  11    ozone    NA
#> 104     8  12    ozone  44.0
#> 105     8  13    ozone  28.0
#> 106     8  14    ozone  65.0
#> 107     8  15    ozone    NA
#> 108     8  16    ozone  22.0
#> 109     8  17    ozone  59.0
#> 110     8  18    ozone  23.0
#> 111     8  19    ozone  31.0
#> 112     8  20    ozone  44.0
#> 113     8  21    ozone  21.0
#> 114     8  22    ozone   9.0
#> 115     8  23    ozone    NA
#> 116     8  24    ozone  45.0
#> 117     8  25    ozone 168.0
#> 118     8  26    ozone  73.0
#> 119     8  27    ozone    NA
#> 120     8  28    ozone  76.0
#> 121     8  29    ozone 118.0
#> 122     8  30    ozone  84.0
#> 123     8  31    ozone  85.0
#> 124     9   1    ozone  96.0
#> 125     9   2    ozone  78.0
#> 126     9   3    ozone  73.0
#> 127     9   4    ozone  91.0
#> 128     9   5    ozone  47.0
#> 129     9   6    ozone  32.0
#> 130     9   7    ozone  20.0
#> 131     9   8    ozone  23.0
#> 132     9   9    ozone  21.0
#> 133     9  10    ozone  24.0
#> 134     9  11    ozone  44.0
#> 135     9  12    ozone  21.0
#> 136     9  13    ozone  28.0
#> 137     9  14    ozone   9.0
#> 138     9  15    ozone  13.0
#> 139     9  16    ozone  46.0
#> 140     9  17    ozone  18.0
#> 141     9  18    ozone  13.0
#> 142     9  19    ozone  24.0
#> 143     9  20    ozone  16.0
#> 144     9  21    ozone  13.0
#> 145     9  22    ozone  23.0
#> 146     9  23    ozone  36.0
#> 147     9  24    ozone   7.0
#> 148     9  25    ozone  14.0
#> 149     9  26    ozone  30.0
#> 150     9  27    ozone    NA
#> 151     9  28    ozone  14.0
#> 152     9  29    ozone  18.0
#> 153     9  30    ozone  20.0
#> 154     5   1  solar.r 190.0
#> 155     5   2  solar.r 118.0
#> 156     5   3  solar.r 149.0
#> 157     5   4  solar.r 313.0
#> 158     5   5  solar.r    NA
#> 159     5   6  solar.r    NA
#> 160     5   7  solar.r 299.0
#> 161     5   8  solar.r  99.0
#> 162     5   9  solar.r  19.0
#> 163     5  10  solar.r 194.0
#> 164     5  11  solar.r    NA
#> 165     5  12  solar.r 256.0
#> 166     5  13  solar.r 290.0
#> 167     5  14  solar.r 274.0
#> 168     5  15  solar.r  65.0
#> 169     5  16  solar.r 334.0
#> 170     5  17  solar.r 307.0
#> 171     5  18  solar.r  78.0
#> 172     5  19  solar.r 322.0
#> 173     5  20  solar.r  44.0
#> 174     5  21  solar.r   8.0
#> 175     5  22  solar.r 320.0
#> 176     5  23  solar.r  25.0
#> 177     5  24  solar.r  92.0
#> 178     5  25  solar.r  66.0
#> 179     5  26  solar.r 266.0
#> 180     5  27  solar.r    NA
#> 181     5  28  solar.r  13.0
#> 182     5  29  solar.r 252.0
#> 183     5  30  solar.r 223.0
#> 184     5  31  solar.r 279.0
#> 185     6   1  solar.r 286.0
#> 186     6   2  solar.r 287.0
#> 187     6   3  solar.r 242.0
#> 188     6   4  solar.r 186.0
#> 189     6   5  solar.r 220.0
#> 190     6   6  solar.r 264.0
#> 191     6   7  solar.r 127.0
#> 192     6   8  solar.r 273.0
#> 193     6   9  solar.r 291.0
#> 194     6  10  solar.r 323.0
#> 195     6  11  solar.r 259.0
#> 196     6  12  solar.r 250.0
#> 197     6  13  solar.r 148.0
#> 198     6  14  solar.r 332.0
#> 199     6  15  solar.r 322.0
#> 200     6  16  solar.r 191.0
#> 201     6  17  solar.r 284.0
#> 202     6  18  solar.r  37.0
#> 203     6  19  solar.r 120.0
#> 204     6  20  solar.r 137.0
#> 205     6  21  solar.r 150.0
#> 206     6  22  solar.r  59.0
#> 207     6  23  solar.r  91.0
#> 208     6  24  solar.r 250.0
#> 209     6  25  solar.r 135.0
#> 210     6  26  solar.r 127.0
#> 211     6  27  solar.r  47.0
#> 212     6  28  solar.r  98.0
#> 213     6  29  solar.r  31.0
#> 214     6  30  solar.r 138.0
#> 215     7   1  solar.r 269.0
#> 216     7   2  solar.r 248.0
#> 217     7   3  solar.r 236.0
#> 218     7   4  solar.r 101.0
#> 219     7   5  solar.r 175.0
#> 220     7   6  solar.r 314.0
#> 221     7   7  solar.r 276.0
#> 222     7   8  solar.r 267.0
#> 223     7   9  solar.r 272.0
#> 224     7  10  solar.r 175.0
#> 225     7  11  solar.r 139.0
#> 226     7  12  solar.r 264.0
#> 227     7  13  solar.r 175.0
#> 228     7  14  solar.r 291.0
#> 229     7  15  solar.r  48.0
#> 230     7  16  solar.r 260.0
#> 231     7  17  solar.r 274.0
#> 232     7  18  solar.r 285.0
#> 233     7  19  solar.r 187.0
#> 234     7  20  solar.r 220.0
#> 235     7  21  solar.r   7.0
#> 236     7  22  solar.r 258.0
#> 237     7  23  solar.r 295.0
#> 238     7  24  solar.r 294.0
#> 239     7  25  solar.r 223.0
#> 240     7  26  solar.r  81.0
#> 241     7  27  solar.r  82.0
#> 242     7  28  solar.r 213.0
#> 243     7  29  solar.r 275.0
#> 244     7  30  solar.r 253.0
#> 245     7  31  solar.r 254.0
#> 246     8   1  solar.r  83.0
#> 247     8   2  solar.r  24.0
#> 248     8   3  solar.r  77.0
#> 249     8   4  solar.r    NA
#> 250     8   5  solar.r    NA
#> 251     8   6  solar.r    NA
#> 252     8   7  solar.r 255.0
#> 253     8   8  solar.r 229.0
#> 254     8   9  solar.r 207.0
#> 255     8  10  solar.r 222.0
#> 256     8  11  solar.r 137.0
#> 257     8  12  solar.r 192.0
#> 258     8  13  solar.r 273.0
#> 259     8  14  solar.r 157.0
#> 260     8  15  solar.r  64.0
#> 261     8  16  solar.r  71.0
#> 262     8  17  solar.r  51.0
#> 263     8  18  solar.r 115.0
#> 264     8  19  solar.r 244.0
#> 265     8  20  solar.r 190.0
#> 266     8  21  solar.r 259.0
#> 267     8  22  solar.r  36.0
#> 268     8  23  solar.r 255.0
#> 269     8  24  solar.r 212.0
#> 270     8  25  solar.r 238.0
#> 271     8  26  solar.r 215.0
#> 272     8  27  solar.r 153.0
#> 273     8  28  solar.r 203.0
#> 274     8  29  solar.r 225.0
#> 275     8  30  solar.r 237.0
#> 276     8  31  solar.r 188.0
#> 277     9   1  solar.r 167.0
#> 278     9   2  solar.r 197.0
#> 279     9   3  solar.r 183.0
#> 280     9   4  solar.r 189.0
#> 281     9   5  solar.r  95.0
#> 282     9   6  solar.r  92.0
#> 283     9   7  solar.r 252.0
#> 284     9   8  solar.r 220.0
#> 285     9   9  solar.r 230.0
#> 286     9  10  solar.r 259.0
#> 287     9  11  solar.r 236.0
#> 288     9  12  solar.r 259.0
#> 289     9  13  solar.r 238.0
#> 290     9  14  solar.r  24.0
#> 291     9  15  solar.r 112.0
#> 292     9  16  solar.r 237.0
#> 293     9  17  solar.r 224.0
#> 294     9  18  solar.r  27.0
#> 295     9  19  solar.r 238.0
#> 296     9  20  solar.r 201.0
#> 297     9  21  solar.r 238.0
#> 298     9  22  solar.r  14.0
#> 299     9  23  solar.r 139.0
#> 300     9  24  solar.r  49.0
#> 301     9  25  solar.r  20.0
#> 302     9  26  solar.r 193.0
#> 303     9  27  solar.r 145.0
#> 304     9  28  solar.r 191.0
#> 305     9  29  solar.r 131.0
#> 306     9  30  solar.r 223.0
#> 307     5   1     wind   7.4
#> 308     5   2     wind   8.0
#> 309     5   3     wind  12.6
#> 310     5   4     wind  11.5
#> 311     5   5     wind  14.3
#> 312     5   6     wind  14.9
#> 313     5   7     wind   8.6
#> 314     5   8     wind  13.8
#> 315     5   9     wind  20.1
#> 316     5  10     wind   8.6
#> 317     5  11     wind   6.9
#> 318     5  12     wind   9.7
#> 319     5  13     wind   9.2
#> 320     5  14     wind  10.9
#> 321     5  15     wind  13.2
#> 322     5  16     wind  11.5
#> 323     5  17     wind  12.0
#> 324     5  18     wind  18.4
#> 325     5  19     wind  11.5
#> 326     5  20     wind   9.7
#> 327     5  21     wind   9.7
#> 328     5  22     wind  16.6
#> 329     5  23     wind   9.7
#> 330     5  24     wind  12.0
#> 331     5  25     wind  16.6
#> 332     5  26     wind  14.9
#> 333     5  27     wind   8.0
#> 334     5  28     wind  12.0
#> 335     5  29     wind  14.9
#> 336     5  30     wind   5.7
#> 337     5  31     wind   7.4
#> 338     6   1     wind   8.6
#> 339     6   2     wind   9.7
#> 340     6   3     wind  16.1
#> 341     6   4     wind   9.2
#> 342     6   5     wind   8.6
#> 343     6   6     wind  14.3
#> 344     6   7     wind   9.7
#> 345     6   8     wind   6.9
#> 346     6   9     wind  13.8
#> 347     6  10     wind  11.5
#> 348     6  11     wind  10.9
#> 349     6  12     wind   9.2
#> 350     6  13     wind   8.0
#> 351     6  14     wind  13.8
#> 352     6  15     wind  11.5
#> 353     6  16     wind  14.9
#> 354     6  17     wind  20.7
#> 355     6  18     wind   9.2
#> 356     6  19     wind  11.5
#> 357     6  20     wind  10.3
#> 358     6  21     wind   6.3
#> 359     6  22     wind   1.7
#> 360     6  23     wind   4.6
#> 361     6  24     wind   6.3
#> 362     6  25     wind   8.0
#> 363     6  26     wind   8.0
#> 364     6  27     wind  10.3
#> 365     6  28     wind  11.5
#> 366     6  29     wind  14.9
#> 367     6  30     wind   8.0
#> 368     7   1     wind   4.1
#> 369     7   2     wind   9.2
#> 370     7   3     wind   9.2
#> 371     7   4     wind  10.9
#> 372     7   5     wind   4.6
#> 373     7   6     wind  10.9
#> 374     7   7     wind   5.1
#> 375     7   8     wind   6.3
#> 376     7   9     wind   5.7
#> 377     7  10     wind   7.4
#> 378     7  11     wind   8.6
#> 379     7  12     wind  14.3
#> 380     7  13     wind  14.9
#> 381     7  14     wind  14.9
#> 382     7  15     wind  14.3
#> 383     7  16     wind   6.9
#> 384     7  17     wind  10.3
#> 385     7  18     wind   6.3
#> 386     7  19     wind   5.1
#> 387     7  20     wind  11.5
#> 388     7  21     wind   6.9
#> 389     7  22     wind   9.7
#> 390     7  23     wind  11.5
#> 391     7  24     wind   8.6
#> 392     7  25     wind   8.0
#> 393     7  26     wind   8.6
#> 394     7  27     wind  12.0
#> 395     7  28     wind   7.4
#> 396     7  29     wind   7.4
#> 397     7  30     wind   7.4
#> 398     7  31     wind   9.2
#> 399     8   1     wind   6.9
#> 400     8   2     wind  13.8
#> 401     8   3     wind   7.4
#> 402     8   4     wind   6.9
#> 403     8   5     wind   7.4
#> 404     8   6     wind   4.6
#> 405     8   7     wind   4.0
#> 406     8   8     wind  10.3
#> 407     8   9     wind   8.0
#> 408     8  10     wind   8.6
#> 409     8  11     wind  11.5
#> 410     8  12     wind  11.5
#> 411     8  13     wind  11.5
#> 412     8  14     wind   9.7
#> 413     8  15     wind  11.5
#> 414     8  16     wind  10.3
#> 415     8  17     wind   6.3
#> 416     8  18     wind   7.4
#> 417     8  19     wind  10.9
#> 418     8  20     wind  10.3
#> 419     8  21     wind  15.5
#> 420     8  22     wind  14.3
#> 421     8  23     wind  12.6
#> 422     8  24     wind   9.7
#> 423     8  25     wind   3.4
#> 424     8  26     wind   8.0
#> 425     8  27     wind   5.7
#> 426     8  28     wind   9.7
#> 427     8  29     wind   2.3
#> 428     8  30     wind   6.3
#> 429     8  31     wind   6.3
#> 430     9   1     wind   6.9
#> 431     9   2     wind   5.1
#> 432     9   3     wind   2.8
#> 433     9   4     wind   4.6
#> 434     9   5     wind   7.4
#> 435     9   6     wind  15.5
#> 436     9   7     wind  10.9
#> 437     9   8     wind  10.3
#> 438     9   9     wind  10.9
#> 439     9  10     wind   9.7
#> 440     9  11     wind  14.9
#> 441     9  12     wind  15.5
#> 442     9  13     wind   6.3
#> 443     9  14     wind  10.9
#> 444     9  15     wind  11.5
#> 445     9  16     wind   6.9
#> 446     9  17     wind  13.8
#> 447     9  18     wind  10.3
#> 448     9  19     wind  10.3
#> 449     9  20     wind   8.0
#> 450     9  21     wind  12.6
#> 451     9  22     wind   9.2
#> 452     9  23     wind  10.3
#> 453     9  24     wind  10.3
#> 454     9  25     wind  16.6
#> 455     9  26     wind   6.9
#> 456     9  27     wind  13.2
#> 457     9  28     wind  14.3
#> 458     9  29     wind   8.0
#> 459     9  30     wind  11.5
#> 460     5   1     temp  67.0
#> 461     5   2     temp  72.0
#> 462     5   3     temp  74.0
#> 463     5   4     temp  62.0
#> 464     5   5     temp  56.0
#> 465     5   6     temp  66.0
#> 466     5   7     temp  65.0
#> 467     5   8     temp  59.0
#> 468     5   9     temp  61.0
#> 469     5  10     temp  69.0
#> 470     5  11     temp  74.0
#> 471     5  12     temp  69.0
#> 472     5  13     temp  66.0
#> 473     5  14     temp  68.0
#> 474     5  15     temp  58.0
#> 475     5  16     temp  64.0
#> 476     5  17     temp  66.0
#> 477     5  18     temp  57.0
#> 478     5  19     temp  68.0
#> 479     5  20     temp  62.0
#> 480     5  21     temp  59.0
#> 481     5  22     temp  73.0
#> 482     5  23     temp  61.0
#> 483     5  24     temp  61.0
#> 484     5  25     temp  57.0
#> 485     5  26     temp  58.0
#> 486     5  27     temp  57.0
#> 487     5  28     temp  67.0
#> 488     5  29     temp  81.0
#> 489     5  30     temp  79.0
#> 490     5  31     temp  76.0
#> 491     6   1     temp  78.0
#> 492     6   2     temp  74.0
#> 493     6   3     temp  67.0
#> 494     6   4     temp  84.0
#> 495     6   5     temp  85.0
#> 496     6   6     temp  79.0
#> 497     6   7     temp  82.0
#> 498     6   8     temp  87.0
#> 499     6   9     temp  90.0
#> 500     6  10     temp  87.0
#> 501     6  11     temp  93.0
#> 502     6  12     temp  92.0
#> 503     6  13     temp  82.0
#> 504     6  14     temp  80.0
#> 505     6  15     temp  79.0
#> 506     6  16     temp  77.0
#> 507     6  17     temp  72.0
#> 508     6  18     temp  65.0
#> 509     6  19     temp  73.0
#> 510     6  20     temp  76.0
#> 511     6  21     temp  77.0
#> 512     6  22     temp  76.0
#> 513     6  23     temp  76.0
#> 514     6  24     temp  76.0
#> 515     6  25     temp  75.0
#> 516     6  26     temp  78.0
#> 517     6  27     temp  73.0
#> 518     6  28     temp  80.0
#> 519     6  29     temp  77.0
#> 520     6  30     temp  83.0
#> 521     7   1     temp  84.0
#> 522     7   2     temp  85.0
#> 523     7   3     temp  81.0
#> 524     7   4     temp  84.0
#> 525     7   5     temp  83.0
#> 526     7   6     temp  83.0
#> 527     7   7     temp  88.0
#> 528     7   8     temp  92.0
#> 529     7   9     temp  92.0
#> 530     7  10     temp  89.0
#> 531     7  11     temp  82.0
#> 532     7  12     temp  73.0
#> 533     7  13     temp  81.0
#> 534     7  14     temp  91.0
#> 535     7  15     temp  80.0
#> 536     7  16     temp  81.0
#> 537     7  17     temp  82.0
#> 538     7  18     temp  84.0
#> 539     7  19     temp  87.0
#> 540     7  20     temp  85.0
#> 541     7  21     temp  74.0
#> 542     7  22     temp  81.0
#> 543     7  23     temp  82.0
#> 544     7  24     temp  86.0
#> 545     7  25     temp  85.0
#> 546     7  26     temp  82.0
#> 547     7  27     temp  86.0
#> 548     7  28     temp  88.0
#> 549     7  29     temp  86.0
#> 550     7  30     temp  83.0
#> 551     7  31     temp  81.0
#> 552     8   1     temp  81.0
#> 553     8   2     temp  81.0
#> 554     8   3     temp  82.0
#> 555     8   4     temp  86.0
#> 556     8   5     temp  85.0
#> 557     8   6     temp  87.0
#> 558     8   7     temp  89.0
#> 559     8   8     temp  90.0
#> 560     8   9     temp  90.0
#> 561     8  10     temp  92.0
#> 562     8  11     temp  86.0
#> 563     8  12     temp  86.0
#> 564     8  13     temp  82.0
#> 565     8  14     temp  80.0
#> 566     8  15     temp  79.0
#> 567     8  16     temp  77.0
#> 568     8  17     temp  79.0
#> 569     8  18     temp  76.0
#> 570     8  19     temp  78.0
#> 571     8  20     temp  78.0
#> 572     8  21     temp  77.0
#> 573     8  22     temp  72.0
#> 574     8  23     temp  75.0
#> 575     8  24     temp  79.0
#> 576     8  25     temp  81.0
#> 577     8  26     temp  86.0
#> 578     8  27     temp  88.0
#> 579     8  28     temp  97.0
#> 580     8  29     temp  94.0
#> 581     8  30     temp  96.0
#> 582     8  31     temp  94.0
#> 583     9   1     temp  91.0
#> 584     9   2     temp  92.0
#> 585     9   3     temp  93.0
#> 586     9   4     temp  93.0
#> 587     9   5     temp  87.0
#> 588     9   6     temp  84.0
#> 589     9   7     temp  80.0
#> 590     9   8     temp  78.0
#> 591     9   9     temp  75.0
#> 592     9  10     temp  73.0
#> 593     9  11     temp  81.0
#> 594     9  12     temp  76.0
#> 595     9  13     temp  77.0
#> 596     9  14     temp  71.0
#> 597     9  15     temp  71.0
#> 598     9  16     temp  78.0
#> 599     9  17     temp  67.0
#> 600     9  18     temp  76.0
#> 601     9  19     temp  68.0
#> 602     9  20     temp  82.0
#> 603     9  21     temp  64.0
#> 604     9  22     temp  71.0
#> 605     9  23     temp  81.0
#> 606     9  24     temp  69.0
#> 607     9  25     temp  63.0
#> 608     9  26     temp  70.0
#> 609     9  27     temp  77.0
#> 610     9  28     temp  75.0
#> 611     9  29     temp  76.0
#> 612     9  30     temp  68.0
names(ChickWeight) <- tolower(names(ChickWeight))
melt(ChickWeight, id=2:4)
#>     time chick diet variable value
#> 1      0     1    1   weight    42
#> 2      2     1    1   weight    51
#> 3      4     1    1   weight    59
#> 4      6     1    1   weight    64
#> 5      8     1    1   weight    76
#> 6     10     1    1   weight    93
#> 7     12     1    1   weight   106
#> 8     14     1    1   weight   125
#> 9     16     1    1   weight   149
#> 10    18     1    1   weight   171
#> 11    20     1    1   weight   199
#> 12    21     1    1   weight   205
#> 13     0     2    1   weight    40
#> 14     2     2    1   weight    49
#> 15     4     2    1   weight    58
#> 16     6     2    1   weight    72
#> 17     8     2    1   weight    84
#> 18    10     2    1   weight   103
#> 19    12     2    1   weight   122
#> 20    14     2    1   weight   138
#> 21    16     2    1   weight   162
#> 22    18     2    1   weight   187
#> 23    20     2    1   weight   209
#> 24    21     2    1   weight   215
#> 25     0     3    1   weight    43
#> 26     2     3    1   weight    39
#> 27     4     3    1   weight    55
#> 28     6     3    1   weight    67
#> 29     8     3    1   weight    84
#> 30    10     3    1   weight    99
#> 31    12     3    1   weight   115
#> 32    14     3    1   weight   138
#> 33    16     3    1   weight   163
#> 34    18     3    1   weight   187
#> 35    20     3    1   weight   198
#> 36    21     3    1   weight   202
#> 37     0     4    1   weight    42
#> 38     2     4    1   weight    49
#> 39     4     4    1   weight    56
#> 40     6     4    1   weight    67
#> 41     8     4    1   weight    74
#> 42    10     4    1   weight    87
#> 43    12     4    1   weight   102
#> 44    14     4    1   weight   108
#> 45    16     4    1   weight   136
#> 46    18     4    1   weight   154
#> 47    20     4    1   weight   160
#> 48    21     4    1   weight   157
#> 49     0     5    1   weight    41
#> 50     2     5    1   weight    42
#> 51     4     5    1   weight    48
#> 52     6     5    1   weight    60
#> 53     8     5    1   weight    79
#> 54    10     5    1   weight   106
#> 55    12     5    1   weight   141
#> 56    14     5    1   weight   164
#> 57    16     5    1   weight   197
#> 58    18     5    1   weight   199
#> 59    20     5    1   weight   220
#> 60    21     5    1   weight   223
#> 61     0     6    1   weight    41
#> 62     2     6    1   weight    49
#> 63     4     6    1   weight    59
#> 64     6     6    1   weight    74
#> 65     8     6    1   weight    97
#> 66    10     6    1   weight   124
#> 67    12     6    1   weight   141
#> 68    14     6    1   weight   148
#> 69    16     6    1   weight   155
#> 70    18     6    1   weight   160
#> 71    20     6    1   weight   160
#> 72    21     6    1   weight   157
#> 73     0     7    1   weight    41
#> 74     2     7    1   weight    49
#> 75     4     7    1   weight    57
#> 76     6     7    1   weight    71
#> 77     8     7    1   weight    89
#> 78    10     7    1   weight   112
#> 79    12     7    1   weight   146
#> 80    14     7    1   weight   174
#> 81    16     7    1   weight   218
#> 82    18     7    1   weight   250
#> 83    20     7    1   weight   288
#> 84    21     7    1   weight   305
#> 85     0     8    1   weight    42
#> 86     2     8    1   weight    50
#> 87     4     8    1   weight    61
#> 88     6     8    1   weight    71
#> 89     8     8    1   weight    84
#> 90    10     8    1   weight    93
#> 91    12     8    1   weight   110
#> 92    14     8    1   weight   116
#> 93    16     8    1   weight   126
#> 94    18     8    1   weight   134
#> 95    20     8    1   weight   125
#> 96     0     9    1   weight    42
#> 97     2     9    1   weight    51
#> 98     4     9    1   weight    59
#> 99     6     9    1   weight    68
#> 100    8     9    1   weight    85
#> 101   10     9    1   weight    96
#> 102   12     9    1   weight    90
#> 103   14     9    1   weight    92
#> 104   16     9    1   weight    93
#> 105   18     9    1   weight   100
#> 106   20     9    1   weight   100
#> 107   21     9    1   weight    98
#> 108    0    10    1   weight    41
#> 109    2    10    1   weight    44
#> 110    4    10    1   weight    52
#> 111    6    10    1   weight    63
#> 112    8    10    1   weight    74
#> 113   10    10    1   weight    81
#> 114   12    10    1   weight    89
#> 115   14    10    1   weight    96
#> 116   16    10    1   weight   101
#> 117   18    10    1   weight   112
#> 118   20    10    1   weight   120
#> 119   21    10    1   weight   124
#> 120    0    11    1   weight    43
#> 121    2    11    1   weight    51
#> 122    4    11    1   weight    63
#> 123    6    11    1   weight    84
#> 124    8    11    1   weight   112
#> 125   10    11    1   weight   139
#> 126   12    11    1   weight   168
#> 127   14    11    1   weight   177
#> 128   16    11    1   weight   182
#> 129   18    11    1   weight   184
#> 130   20    11    1   weight   181
#> 131   21    11    1   weight   175
#> 132    0    12    1   weight    41
#> 133    2    12    1   weight    49
#> 134    4    12    1   weight    56
#> 135    6    12    1   weight    62
#> 136    8    12    1   weight    72
#> 137   10    12    1   weight    88
#> 138   12    12    1   weight   119
#> 139   14    12    1   weight   135
#> 140   16    12    1   weight   162
#> 141   18    12    1   weight   185
#> 142   20    12    1   weight   195
#> 143   21    12    1   weight   205
#> 144    0    13    1   weight    41
#> 145    2    13    1   weight    48
#> 146    4    13    1   weight    53
#> 147    6    13    1   weight    60
#> 148    8    13    1   weight    65
#> 149   10    13    1   weight    67
#> 150   12    13    1   weight    71
#> 151   14    13    1   weight    70
#> 152   16    13    1   weight    71
#> 153   18    13    1   weight    81
#> 154   20    13    1   weight    91
#> 155   21    13    1   weight    96
#> 156    0    14    1   weight    41
#> 157    2    14    1   weight    49
#> 158    4    14    1   weight    62
#> 159    6    14    1   weight    79
#> 160    8    14    1   weight   101
#> 161   10    14    1   weight   128
#> 162   12    14    1   weight   164
#> 163   14    14    1   weight   192
#> 164   16    14    1   weight   227
#> 165   18    14    1   weight   248
#> 166   20    14    1   weight   259
#> 167   21    14    1   weight   266
#> 168    0    15    1   weight    41
#> 169    2    15    1   weight    49
#> 170    4    15    1   weight    56
#> 171    6    15    1   weight    64
#> 172    8    15    1   weight    68
#> 173   10    15    1   weight    68
#> 174   12    15    1   weight    67
#> 175   14    15    1   weight    68
#> 176    0    16    1   weight    41
#> 177    2    16    1   weight    45
#> 178    4    16    1   weight    49
#> 179    6    16    1   weight    51
#> 180    8    16    1   weight    57
#> 181   10    16    1   weight    51
#> 182   12    16    1   weight    54
#> 183    0    17    1   weight    42
#> 184    2    17    1   weight    51
#> 185    4    17    1   weight    61
#> 186    6    17    1   weight    72
#> 187    8    17    1   weight    83
#> 188   10    17    1   weight    89
#> 189   12    17    1   weight    98
#> 190   14    17    1   weight   103
#> 191   16    17    1   weight   113
#> 192   18    17    1   weight   123
#> 193   20    17    1   weight   133
#> 194   21    17    1   weight   142
#> 195    0    18    1   weight    39
#> 196    2    18    1   weight    35
#> 197    0    19    1   weight    43
#> 198    2    19    1   weight    48
#> 199    4    19    1   weight    55
#> 200    6    19    1   weight    62
#> 201    8    19    1   weight    65
#> 202   10    19    1   weight    71
#> 203   12    19    1   weight    82
#> 204   14    19    1   weight    88
#> 205   16    19    1   weight   106
#> 206   18    19    1   weight   120
#> 207   20    19    1   weight   144
#> 208   21    19    1   weight   157
#> 209    0    20    1   weight    41
#> 210    2    20    1   weight    47
#> 211    4    20    1   weight    54
#> 212    6    20    1   weight    58
#> 213    8    20    1   weight    65
#> 214   10    20    1   weight    73
#> 215   12    20    1   weight    77
#> 216   14    20    1   weight    89
#> 217   16    20    1   weight    98
#> 218   18    20    1   weight   107
#> 219   20    20    1   weight   115
#> 220   21    20    1   weight   117
#> 221    0    21    2   weight    40
#> 222    2    21    2   weight    50
#> 223    4    21    2   weight    62
#> 224    6    21    2   weight    86
#> 225    8    21    2   weight   125
#> 226   10    21    2   weight   163
#> 227   12    21    2   weight   217
#> 228   14    21    2   weight   240
#> 229   16    21    2   weight   275
#> 230   18    21    2   weight   307
#> 231   20    21    2   weight   318
#> 232   21    21    2   weight   331
#> 233    0    22    2   weight    41
#> 234    2    22    2   weight    55
#> 235    4    22    2   weight    64
#> 236    6    22    2   weight    77
#> 237    8    22    2   weight    90
#> 238   10    22    2   weight    95
#> 239   12    22    2   weight   108
#> 240   14    22    2   weight   111
#> 241   16    22    2   weight   131
#> 242   18    22    2   weight   148
#> 243   20    22    2   weight   164
#> 244   21    22    2   weight   167
#> 245    0    23    2   weight    43
#> 246    2    23    2   weight    52
#> 247    4    23    2   weight    61
#> 248    6    23    2   weight    73
#> 249    8    23    2   weight    90
#> 250   10    23    2   weight   103
#> 251   12    23    2   weight   127
#> 252   14    23    2   weight   135
#> 253   16    23    2   weight   145
#> 254   18    23    2   weight   163
#> 255   20    23    2   weight   170
#> 256   21    23    2   weight   175
#> 257    0    24    2   weight    42
#> 258    2    24    2   weight    52
#> 259    4    24    2   weight    58
#> 260    6    24    2   weight    74
#> 261    8    24    2   weight    66
#> 262   10    24    2   weight    68
#> 263   12    24    2   weight    70
#> 264   14    24    2   weight    71
#> 265   16    24    2   weight    72
#> 266   18    24    2   weight    72
#> 267   20    24    2   weight    76
#> 268   21    24    2   weight    74
#> 269    0    25    2   weight    40
#> 270    2    25    2   weight    49
#> 271    4    25    2   weight    62
#> 272    6    25    2   weight    78
#> 273    8    25    2   weight   102
#> 274   10    25    2   weight   124
#> 275   12    25    2   weight   146
#> 276   14    25    2   weight   164
#> 277   16    25    2   weight   197
#> 278   18    25    2   weight   231
#> 279   20    25    2   weight   259
#> 280   21    25    2   weight   265
#> 281    0    26    2   weight    42
#> 282    2    26    2   weight    48
#> 283    4    26    2   weight    57
#> 284    6    26    2   weight    74
#> 285    8    26    2   weight    93
#> 286   10    26    2   weight   114
#> 287   12    26    2   weight   136
#> 288   14    26    2   weight   147
#> 289   16    26    2   weight   169
#> 290   18    26    2   weight   205
#> 291   20    26    2   weight   236
#> 292   21    26    2   weight   251
#> 293    0    27    2   weight    39
#> 294    2    27    2   weight    46
#> 295    4    27    2   weight    58
#> 296    6    27    2   weight    73
#> 297    8    27    2   weight    87
#> 298   10    27    2   weight   100
#> 299   12    27    2   weight   115
#> 300   14    27    2   weight   123
#> 301   16    27    2   weight   144
#> 302   18    27    2   weight   163
#> 303   20    27    2   weight   185
#> 304   21    27    2   weight   192
#> 305    0    28    2   weight    39
#> 306    2    28    2   weight    46
#> 307    4    28    2   weight    58
#> 308    6    28    2   weight    73
#> 309    8    28    2   weight    92
#> 310   10    28    2   weight   114
#> 311   12    28    2   weight   145
#> 312   14    28    2   weight   156
#> 313   16    28    2   weight   184
#> 314   18    28    2   weight   207
#> 315   20    28    2   weight   212
#> 316   21    28    2   weight   233
#> 317    0    29    2   weight    39
#> 318    2    29    2   weight    48
#> 319    4    29    2   weight    59
#> 320    6    29    2   weight    74
#> 321    8    29    2   weight    87
#> 322   10    29    2   weight   106
#> 323   12    29    2   weight   134
#> 324   14    29    2   weight   150
#> 325   16    29    2   weight   187
#> 326   18    29    2   weight   230
#> 327   20    29    2   weight   279
#> 328   21    29    2   weight   309
#> 329    0    30    2   weight    42
#> 330    2    30    2   weight    48
#> 331    4    30    2   weight    59
#> 332    6    30    2   weight    72
#> 333    8    30    2   weight    85
#> 334   10    30    2   weight    98
#> 335   12    30    2   weight   115
#> 336   14    30    2   weight   122
#> 337   16    30    2   weight   143
#> 338   18    30    2   weight   151
#> 339   20    30    2   weight   157
#> 340   21    30    2   weight   150
#> 341    0    31    3   weight    42
#> 342    2    31    3   weight    53
#> 343    4    31    3   weight    62
#> 344    6    31    3   weight    73
#> 345    8    31    3   weight    85
#> 346   10    31    3   weight   102
#> 347   12    31    3   weight   123
#> 348   14    31    3   weight   138
#> 349   16    31    3   weight   170
#> 350   18    31    3   weight   204
#> 351   20    31    3   weight   235
#> 352   21    31    3   weight   256
#> 353    0    32    3   weight    41
#> 354    2    32    3   weight    49
#> 355    4    32    3   weight    65
#> 356    6    32    3   weight    82
#> 357    8    32    3   weight   107
#> 358   10    32    3   weight   129
#> 359   12    32    3   weight   159
#> 360   14    32    3   weight   179
#> 361   16    32    3   weight   221
#> 362   18    32    3   weight   263
#> 363   20    32    3   weight   291
#> 364   21    32    3   weight   305
#> 365    0    33    3   weight    39
#> 366    2    33    3   weight    50
#> 367    4    33    3   weight    63
#> 368    6    33    3   weight    77
#> 369    8    33    3   weight    96
#> 370   10    33    3   weight   111
#> 371   12    33    3   weight   137
#> 372   14    33    3   weight   144
#> 373   16    33    3   weight   151
#> 374   18    33    3   weight   146
#> 375   20    33    3   weight   156
#> 376   21    33    3   weight   147
#> 377    0    34    3   weight    41
#> 378    2    34    3   weight    49
#> 379    4    34    3   weight    63
#> 380    6    34    3   weight    85
#> 381    8    34    3   weight   107
#> 382   10    34    3   weight   134
#> 383   12    34    3   weight   164
#> 384   14    34    3   weight   186
#> 385   16    34    3   weight   235
#> 386   18    34    3   weight   294
#> 387   20    34    3   weight   327
#> 388   21    34    3   weight   341
#> 389    0    35    3   weight    41
#> 390    2    35    3   weight    53
#> 391    4    35    3   weight    64
#> 392    6    35    3   weight    87
#> 393    8    35    3   weight   123
#> 394   10    35    3   weight   158
#> 395   12    35    3   weight   201
#> 396   14    35    3   weight   238
#> 397   16    35    3   weight   287
#> 398   18    35    3   weight   332
#> 399   20    35    3   weight   361
#> 400   21    35    3   weight   373
#> 401    0    36    3   weight    39
#> 402    2    36    3   weight    48
#> 403    4    36    3   weight    61
#> 404    6    36    3   weight    76
#> 405    8    36    3   weight    98
#> 406   10    36    3   weight   116
#> 407   12    36    3   weight   145
#> 408   14    36    3   weight   166
#> 409   16    36    3   weight   198
#> 410   18    36    3   weight   227
#> 411   20    36    3   weight   225
#> 412   21    36    3   weight   220
#> 413    0    37    3   weight    41
#> 414    2    37    3   weight    48
#> 415    4    37    3   weight    56
#> 416    6    37    3   weight    68
#> 417    8    37    3   weight    80
#> 418   10    37    3   weight    83
#> 419   12    37    3   weight   103
#> 420   14    37    3   weight   112
#> 421   16    37    3   weight   135
#> 422   18    37    3   weight   157
#> 423   20    37    3   weight   169
#> 424   21    37    3   weight   178
#> 425    0    38    3   weight    41
#> 426    2    38    3   weight    49
#> 427    4    38    3   weight    61
#> 428    6    38    3   weight    74
#> 429    8    38    3   weight    98
#> 430   10    38    3   weight   109
#> 431   12    38    3   weight   128
#> 432   14    38    3   weight   154
#> 433   16    38    3   weight   192
#> 434   18    38    3   weight   232
#> 435   20    38    3   weight   280
#> 436   21    38    3   weight   290
#> 437    0    39    3   weight    42
#> 438    2    39    3   weight    50
#> 439    4    39    3   weight    61
#> 440    6    39    3   weight    78
#> 441    8    39    3   weight    89
#> 442   10    39    3   weight   109
#> 443   12    39    3   weight   130
#> 444   14    39    3   weight   146
#> 445   16    39    3   weight   170
#> 446   18    39    3   weight   214
#> 447   20    39    3   weight   250
#> 448   21    39    3   weight   272
#> 449    0    40    3   weight    41
#> 450    2    40    3   weight    55
#> 451    4    40    3   weight    66
#> 452    6    40    3   weight    79
#> 453    8    40    3   weight   101
#> 454   10    40    3   weight   120
#> 455   12    40    3   weight   154
#> 456   14    40    3   weight   182
#> 457   16    40    3   weight   215
#> 458   18    40    3   weight   262
#> 459   20    40    3   weight   295
#> 460   21    40    3   weight   321
#> 461    0    41    4   weight    42
#> 462    2    41    4   weight    51
#> 463    4    41    4   weight    66
#> 464    6    41    4   weight    85
#> 465    8    41    4   weight   103
#> 466   10    41    4   weight   124
#> 467   12    41    4   weight   155
#> 468   14    41    4   weight   153
#> 469   16    41    4   weight   175
#> 470   18    41    4   weight   184
#> 471   20    41    4   weight   199
#> 472   21    41    4   weight   204
#> 473    0    42    4   weight    42
#> 474    2    42    4   weight    49
#> 475    4    42    4   weight    63
#> 476    6    42    4   weight    84
#> 477    8    42    4   weight   103
#> 478   10    42    4   weight   126
#> 479   12    42    4   weight   160
#> 480   14    42    4   weight   174
#> 481   16    42    4   weight   204
#> 482   18    42    4   weight   234
#> 483   20    42    4   weight   269
#> 484   21    42    4   weight   281
#> 485    0    43    4   weight    42
#> 486    2    43    4   weight    55
#> 487    4    43    4   weight    69
#> 488    6    43    4   weight    96
#> 489    8    43    4   weight   131
#> 490   10    43    4   weight   157
#> 491   12    43    4   weight   184
#> 492   14    43    4   weight   188
#> 493   16    43    4   weight   197
#> 494   18    43    4   weight   198
#> 495   20    43    4   weight   199
#> 496   21    43    4   weight   200
#> 497    0    44    4   weight    42
#> 498    2    44    4   weight    51
#> 499    4    44    4   weight    65
#> 500    6    44    4   weight    86
#> 501    8    44    4   weight   103
#> 502   10    44    4   weight   118
#> 503   12    44    4   weight   127
#> 504   14    44    4   weight   138
#> 505   16    44    4   weight   145
#> 506   18    44    4   weight   146
#> 507    0    45    4   weight    41
#> 508    2    45    4   weight    50
#> 509    4    45    4   weight    61
#> 510    6    45    4   weight    78
#> 511    8    45    4   weight    98
#> 512   10    45    4   weight   117
#> 513   12    45    4   weight   135
#> 514   14    45    4   weight   141
#> 515   16    45    4   weight   147
#> 516   18    45    4   weight   174
#> 517   20    45    4   weight   197
#> 518   21    45    4   weight   196
#> 519    0    46    4   weight    40
#> 520    2    46    4   weight    52
#> 521    4    46    4   weight    62
#> 522    6    46    4   weight    82
#> 523    8    46    4   weight   101
#> 524   10    46    4   weight   120
#> 525   12    46    4   weight   144
#> 526   14    46    4   weight   156
#> 527   16    46    4   weight   173
#> 528   18    46    4   weight   210
#> 529   20    46    4   weight   231
#> 530   21    46    4   weight   238
#> 531    0    47    4   weight    41
#> 532    2    47    4   weight    53
#> 533    4    47    4   weight    66
#> 534    6    47    4   weight    79
#> 535    8    47    4   weight   100
#> 536   10    47    4   weight   123
#> 537   12    47    4   weight   148
#> 538   14    47    4   weight   157
#> 539   16    47    4   weight   168
#> 540   18    47    4   weight   185
#> 541   20    47    4   weight   210
#> 542   21    47    4   weight   205
#> 543    0    48    4   weight    39
#> 544    2    48    4   weight    50
#> 545    4    48    4   weight    62
#> 546    6    48    4   weight    80
#> 547    8    48    4   weight   104
#> 548   10    48    4   weight   125
#> 549   12    48    4   weight   154
#> 550   14    48    4   weight   170
#> 551   16    48    4   weight   222
#> 552   18    48    4   weight   261
#> 553   20    48    4   weight   303
#> 554   21    48    4   weight   322
#> 555    0    49    4   weight    40
#> 556    2    49    4   weight    53
#> 557    4    49    4   weight    64
#> 558    6    49    4   weight    85
#> 559    8    49    4   weight   108
#> 560   10    49    4   weight   128
#> 561   12    49    4   weight   152
#> 562   14    49    4   weight   166
#> 563   16    49    4   weight   184
#> 564   18    49    4   weight   203
#> 565   20    49    4   weight   233
#> 566   21    49    4   weight   237
#> 567    0    50    4   weight    41
#> 568    2    50    4   weight    54
#> 569    4    50    4   weight    67
#> 570    6    50    4   weight    84
#> 571    8    50    4   weight   105
#> 572   10    50    4   weight   122
#> 573   12    50    4   weight   155
#> 574   14    50    4   weight   175
#> 575   16    50    4   weight   205
#> 576   18    50    4   weight   234
#> 577   20    50    4   weight   264
#> 578   21    50    4   weight   264