melt.default {reshape2}	R Documentation

Melt a vector. For vectors, makes a column of a data frame

Description

Melt a vector. For vectors, makes a column of a data frame

Usage

## Default S3 method:
melt(data, ..., na.rm = FALSE, value.name = "value")

Arguments

data
vector to melt

...
further arguments passed to or from other methods.

na.rm
Should NA values be removed from the data set? This will convert explicit missings to implicit missings.

value.name
name of variable used to store values

See Also

melt, cast

Other melt methods: melt.array(), melt.data.frame(), melt.list()

[Package reshape2 version 1.4.4 Index]